requirements.txt
scikit-learn
matplotlib
pylast
gnureadline
apache-airflow-providers-amazon
apache-airflow-backport-providers-amazon
boto3
